+++ src/index.ts
@@ -7,15 +7,18 @@ export { ButtonTypes } from "./components/Button/ButtonTypes";
 export { default as Card } from "./components/Card/Card";
 export { default as Checkbox } from "./components/Checkbox/Checkbox";
 export { default as Heading } from "./components/Heading/Heading";
+export { HeadingDisplaySizes } from "./components/Heading/HeadingDisplaySizes";
 export { default as HelperErrorText } from "./components/HelperErrorText/HelperErrorText";
 export { default as Hero } from "./components/Hero/Hero";
 export { HeroTypes } from "./components/Hero/HeroTypes";
 export { default as HorizontalRule } from "./components/HorizontalRule/HorizontalRule";
 export { default as Icon } from "./components/Icons/Icon";
 export {
-  IconRotationTypes,
   IconNames,
   LogoNames,
+  IconRotationTypes,
+  IconColors,
+  IconSizes,
 } from "./components/Icons/IconTypes";
 export { default as Image } from "./components/Image/Image";
 export { default as Input } from "./components/Input/Input";